create tools directory
It contains a set of tools (VMs, compilers, debuggers) for
bareboard development together with configuration scripts.

The purpose is to have a simple way to recreate the development
environment for a new machine in case this is needed.

Usually some tools should be installed through the command line
by using different package managers. When this is not available
the tool can be compiled and configured from source code.

Currently Windows and Linux are the only two environments targeted.

Signed-off-by: Radu Alexe <alexeradu2007@gmail.com>