init Vagrant box
Having a different setup for Windows/Linux has proven to be hard.
I am following currently a solution where I install a VMBox machine
via vagrant and configure everything inside it.

This should be easier and provide a single configuration setup
for both Linux and Windows(on VMBox).

Linux could use a VMBox as well.