add function to read sectors from memory
It seems that it is too large to fit into the bootsector. The problem is the
divisions to weird numbers (like 18 or 80) which the compiler tries to break
them into other instructions.

The solution would be to re-write (or partially write) this function in
assembly with the inspiration from boot/disk.asm.

This would be the last function that should exist in the bootsector!!!

Signed-off-by: Radu Alexe <radu.alexe@nxp.com>